Refactor Go Pixie API query code from Grafana query handling
