+++ pkg/pixie_plugin.go
@@ -22,49 +22,20 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
-	"strings"
-	"time"
 
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/datasource"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
-	"github.com/grafana/grafana-plugin-sdk-go/data"
 
 	"px.dev/pxapi"
 )
-
-// GrafanaMacro is a type which defines a macro.
-type GrafanaMacro string
 
 const (
 	// Define keys to retrieve configs passed from UI.
 	apiKeyField    = "apiKey"
 	cloudAddrField = "cloudAddr"
 	clusterIDField = "clusterId"
-	// timeFromMacro is the start of the time range of a query.
-	timeFromMacro GrafanaMacro = "__time_from"
-	// timeToMacro is the end of the time range of a query.
-	timeToMacro GrafanaMacro = "__time_to"
-	// intervalMacro is the suggested duration between time points.
-	intervalMacro GrafanaMacro = "__interval"
 )
-
-// replaceTimeMacroInQueryText takes the query text (PxL script to execute)
-// and replaces the time macros with the relevant time objects.
-func replaceTimeMacroInQueryText(queryText string, grafanaMacro GrafanaMacro,
-	timeReplacement time.Time) string {
-	tStr := fmt.Sprintf("%d", timeReplacement.UnixNano())
-	return strings.ReplaceAll(queryText, string(grafanaMacro), tStr)
-}
-
-// replaceIntervalMacroInQueryText takes the query text and replaces
-// interval macro with interval duration specified in Grafana UI.
-func replaceIntervalMacroInQueryText(queryText string, grafanaMacro GrafanaMacro,
-	intervalDuration time.Duration) string {
-	intervalStr := fmt.Sprintf("%d", intervalDuration.Nanoseconds())
-	return strings.ReplaceAll(queryText, string(grafanaMacro), intervalStr)
-}
 
 // createPixieDatasource creates a new Pixie datasource.
 func createPixieDatasource() datasource.ServeOpts {
@@ -103,7 +74,7 @@ type queryModel struct {
 	ClusterFlag bool   `json:bool`
 }
 
-func createVizierClient(ctx context.Context, apiKey string, clusterID string, cloudAddr string) (*pxapi.VizierClient, error) {
+func createClient(ctx context.Context, apiKey string, cloudAddr string) (*pxapi.Client, error) {
 	var client *pxapi.Client
 	var err error
 	// First, create a client connecting to Pixie Cloud.
@@ -115,17 +86,16 @@ func createVizierClient(ctx context.Context, apiKey string, clusterID string, cl
 	if err != nil {
 		return nil, err
 	}
-	// Next, create a client that connects to the particular Vizier instance matching `clusterID`.
-	vzClient, err := client.NewVizierClient(ctx, clusterID)
 
-	if err != nil {
-		return nil, err
-	}
-	return vzClient, nil
+	return client, nil
 }
 
 func (td *PixieDatasource) query(ctx context.Context, query backend.DataQuery,
 	config map[string]string) (*backend.DataResponse, error) {
+
+	apiToken := config[apiKeyField]
+	clusterID := config[clusterIDField]
+	cloudAddr := config[cloudAddrField]
 
 	var qm queryModel
 	err := json.Unmarshal(query.JSON, &qm)
@@ -133,109 +103,16 @@ func (td *PixieDatasource) query(ctx context.Context, query backend.DataQuery,
 		return nil, fmt.Errorf("Error unmarshalling JSON: %v", err)
 	}
 
-	apiToken := config[apiKeyField]
-	clusterID := config[clusterIDField]
-	cloudAddr := config[cloudAddrField]
+	client, err := createClient(ctx, apiToken, cloudAddr)
+	qp := PixieQueryProcessor{
+		client: client,
+	}
 
 	if qm.ClusterFlag {
-		return td.queryClusters(ctx, apiToken)
+		return qp.queryClusters(ctx, apiToken)
+	} else {
+		return qp.queryScript(ctx, qm, query, clusterID)
 	}
-
-	vz, err := createVizierClient(ctx, apiToken, clusterID, cloudAddr)
-	if err != nil {
-		log.DefaultLogger.Error(fmt.Sprintf("Unable to create Vizier Client: %+v", err))
-		return nil, err
-	}
-
-	response := &backend.DataResponse{}
-
-	// Create TableMuxer to accept results table.
-	tm := &PixieToGrafanaTableMux{}
-
-	// Update macros in query text.
-	qm.PxlScript = replaceTimeMacroInQueryText(qm.PxlScript, timeFromMacro,
-		query.TimeRange.From)
-	qm.PxlScript = replaceTimeMacroInQueryText(qm.PxlScript, timeToMacro,
-		query.TimeRange.To)
-	qm.PxlScript = replaceIntervalMacroInQueryText(qm.PxlScript, intervalMacro,
-		query.Interval)
-
-	// Execute the PxL script.
-	resultSet, err := vz.ExecuteScript(ctx, qm.PxlScript, tm)
-	if err != nil && err != io.EOF {
-		log.DefaultLogger.Warn("Can't execute script.")
-		return nil, err
-	}
-
-	// Receive the PxL script results.
-	defer resultSet.Close()
-	if err := resultSet.Stream(); err != nil {
-		streamStrErr := fmt.Errorf("got error : %+v, while streaming", err)
-		response.Error = streamStrErr
-		log.DefaultLogger.Error(streamStrErr.Error())
-	}
-
-	// Add the frames to the response.
-	for _, tablePrinter := range tm.pxTablePrinterLst {
-		// If time series schema long && time_ column, convert to wide. Otherwise
-		// proceed as normal.
-		tsSchema := tablePrinter.frame.TimeSeriesSchema()
-		if tablePrinter.FormatGrafanaTimeFrame() && tsSchema.Type == data.TimeSeriesTypeLong {
-			wideFrame, err := data.LongToWide(tablePrinter.frame,
-				&data.FillMissing{Mode: data.FillModeNull})
-			if err != nil {
-				return nil, err
-			}
-			response.Frames = append(response.Frames, wideFrame)
-		} else {
-			response.Frames = append(response.Frames,
-				tablePrinter.frame)
-		}
-	}
-	return response, nil
-}
-
-func (td *PixieDatasource) queryClusters(ctx context.Context, apiToken string) (*backend.DataResponse, error) {
-
-	response := &backend.DataResponse{}
-
-	client, err := pxapi.NewClient(ctx, pxapi.WithAPIKey(apiToken))
-	if err != nil {
-		return nil, fmt.Errorf("Error unmarshalling JSON: %v", err)
-	}
-
-	var viziers []*pxapi.VizierInfo
-	viziers, err = client.ListViziers(ctx)
-
-	if err != nil {
-		return nil, fmt.Errorf("Error with getting viziers: $s", err)
-	}
-
-	vizierIds := make([]string, 0)
-	vizierNames := make([]string, 0)
-
-	for _, vizier := range viziers {
-		// Only show connected clusters
-		if vizier.Status != pxapi.VizierStatusDisconnected {
-			vizierIds = append(vizierIds, vizier.ID)
-			vizierNames = append(vizierNames, vizier.Name)
-		}
-	}
-
-	vizierFrame := data.NewFrame(
-		"Vizier Clusters",
-		data.NewField("id", data.Labels{}, vizierIds),
-		data.NewField("name", data.Labels{}, vizierNames),
-	)
-
-	response.Frames = append(response.Frames, vizierFrame)
-
-	backend.Logger.Debug(fmt.Sprintf("number of viziers: $d", len(viziers)))
-	backend.Logger.Debug(fmt.Sprintf("id: $s, name: $s, status: $s", viziers[0].ID, viziers[0].Name, viziers[0].Status))
-
-	backend.Logger.Debug(vizierFrame.StringTable(10, 10))
-
-	return response, nil
 }
 
 // CheckHealth implements the Grafana service health check API.
@@ -244,7 +121,14 @@ func (td *PixieDatasource) CheckHealth(ctx context.Context, req *backend.CheckHe
 	message := "Connection to Pixie cluster successfully configured"
 	config := req.PluginContext.DataSourceInstanceSettings.DecryptedSecureJSONData
 
-	vz, err := createVizierClient(ctx, config[apiKeyField], config[clusterIDField], config[cloudAddrField])
+	client, err := createClient(ctx, config[apiKeyField], config[cloudAddrField])
+	if err != nil {
+		status = backend.HealthStatusError
+		message = fmt.Sprintf("Error connecting Pixie client: %s", err.Error())
+	}
+
+	vz, err := client.NewVizierClient(ctx, config[clusterIDField])
+
 	if vz == nil || err != nil {
 		status = backend.HealthStatusError
 		message = fmt.Sprintf("Error connecting to Vizier: %s", err.Error())
